requirements.txt
keras==3.5.0
tensorflow==2.17.0
PyQt5==5.15.11
opencv-python==4.10.0.84
scikit-learn==1.5.1
numpy==1.26.4
pillow==10.4.0
matplotlib==3.9.2
